Using vec u32 instead of float for px precision

+++ gui/src/common/events.rs
@@ -7,6 +7,6 @@ pub enum WidgetEvent {
     Exiting(UID),
     Pressed(UID),
     Released(UID),
-    Dragging(UID, Vector2f),
+    Dragging(UID, Vector2i),
 }
 impl Event for WidgetEvent {}
